requirements.txt
google-genai
openai
requests
certifi
python-dotenv
cryptography
Pillow